+++ src/main/java/com/mageddo/dnsproxyserver/server/dns/solver/HostnameQuery.java
@@ -2,7 +2,7 @@ package com.mageddo.dnsproxyserver.server.dns.solver;
 
 import com.mageddo.dnsproxyserver.config.Config.Entry.Type;
 import com.mageddo.dns.Hostname;
-import com.mageddo.dns.Wildcards;
+import com.mageddo.dns.utils.Wildcards;
 import com.mageddo.net.IP;
 import lombok.Builder;
 import lombok.EqualsAndHashCode;
