+++ src/main/java/com/mageddo/dnsproxyserver/server/dns/solver/SolverRemote.java
@@ -4,7 +4,7 @@ import com.mageddo.commons.circuitbreaker.CircuitCheckException;
 import com.mageddo.commons.concurrent.ThreadPool;
 import com.mageddo.commons.concurrent.Threads;
 import com.mageddo.dnsproxyserver.config.application.ConfigService;
-import com.mageddo.dns.Messages;
+import com.mageddo.dns.utils.Messages;
 import com.mageddo.net.Networks;
 import dev.failsafe.CircuitBreaker;
 import dev.failsafe.CircuitBreakerOpenException;
@@ -29,7 +29,7 @@ import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Future;
 import java.util.concurrent.atomic.AtomicReference;
 
-import static com.mageddo.dns.Messages.simplePrint;
+import static com.mageddo.dns.utils.Messages.simplePrint;
 
 @Slf4j
 @Singleton
