+++ src/test/java/com/mageddo/dnsproxyserver/dns/server/solver/WildcardsTest.java
@@ -3,7 +3,7 @@ package com.mageddo.dnsproxyserver.dns.server.solver;
 import com.mageddo.dns.Hostname;
 import org.junit.jupiter.api.Test;
 
-import static com.mageddo.dns.Wildcards.buildHostAndWildcards;
+import static com.mageddo.dns.utils.Wildcards.buildHostAndWildcards;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 class WildcardsTest {
