+++ src/test/java/com/mageddo/dnsproxyserver/server/dns/MessagesTest.java
@@ -1,12 +1,12 @@
 package com.mageddo.dnsproxyserver.server.dns;
 
-import com.mageddo.dns.Messages;
+import com.mageddo.dns.utils.Messages;
 import testing.templates.MessageTemplates;
 import org.junit.jupiter.api.Test;
 import org.xbill.DNS.Flags;
 import org.xbill.DNS.Rcode;
 
-import static com.mageddo.dns.Messages.findFirstAnswerRecord;
+import static com.mageddo.dns.utils.Messages.findFirstAnswerRecord;
 import static testing.templates.MessageTemplates.acmeAResponse;
 import static testing.templates.MessageTemplates.acmeNxDomain;
 import static org.junit.jupiter.api.Assertions.assertEquals;
